Moved some button variant CSS to main.css because we have the slideshow variant css here too.
Like slideshow variants, the button variant is a nice example for how to take advantage of this functionality, but we don't need to include it in the plugin, just the sandbox.

We often override these styles to suit our needs anyway and it makes sense to have them in this location.




git-svn-id: http://svn.apostrophenow.org/sandboxes/asandbox/trunk@2156 c89fab08-2f9b-42bf-8176-bd3ad6bee547